requirements.txt
paho-mqtt==1.6.1
python-dotenv==1.0.0
pydantic==2.5.0
schedule==1.2.0
